Update the slider properly
